requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
redis
rq
requests
openai
python-jose
pydantic
pydantic[email]
